second commit

+++ src/main/java/com/example/voltix/Zones/ZoneService.java
@@ -1,6 +1,7 @@
 package com.example.voltix.Zones;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -12,7 +13,7 @@ public class ZoneService {
     @Autowired
     private ZoneRepository zoneRepository;
 
-        public ZoneModel saveOrUpdate(ZoneModel zone){
+        public ZoneModel addZone(ZoneModel zone){
             return zoneRepository.save(zone);
         }
 
@@ -20,12 +21,24 @@ public class ZoneService {
             return zoneRepository.findAll();
         }
 
-        public void delete(ZoneModel zone) {
-            zoneRepository.delete(zone);
+        public ZoneModel findZoneByName(String zoneName) {
+            return zoneRepository.findByZoneName(zoneName);
         }
 
+        public void deleteZoneById(String id) {
+            Optional<ZoneModel> zoneOptional = zoneRepository.findById(id);
+            if (zoneOptional.isPresent()) {
+                ZoneModel zone = zoneOptional.get();
+                zoneRepository.delete(zone);
+            } else {
+                // Handle the case when the zone with the specified ID does not exist.
+                // You can throw an exception, log a message, or take any other appropriate action.
+            }
+        }
 
-
+        public ZoneModel findZoneById(String zoneId) {
+            return zoneRepository.findById(zoneId).orElse(null);
+        }
